requirements.txt
hdfs
python-decouple
